Gate all document + booking routes behind GrowthBook flags, rename booking-events to booking-preorders
- documentsEnabled check on /documents, /documents/[slug], /buy-documents, /buy-documents/[slug]
- bookingsEnabled check on /bookings, /bookings/[slug], /book/[slug], /account/booking-preorders
- Rename booking-events → booking-preorders (client + admin routes, nav, all internal links)
- Add test item guard to document guide page
